Chore: Exception for services, custom exception for DTO, and global exception in controllers that handle requests

+++ ecommerce/src/main/java/com/campos/ecommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.campos.ecommerce.services;
 import com.campos.ecommerce.dto.ProductDto;
 import com.campos.ecommerce.entities.Product;
 import com.campos.ecommerce.repositories.ProductRepository;
+import com.campos.ecommerce.services.exceptions.DatabaseException;
+import com.campos.ecommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
@@ -20,12 +25,12 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDto findById(Long id) {
         /*
-        * Product product = repository.findById(id).get();
+        * Product product = productRepository.findById(id).get();
         * return new ProductDto(product);
         * */
         Optional<Product> result = productRepository.findById(id);
         System.out.println("result: " + result);
-        Product product = result.get();
+        Product product = result.orElseThrow(() -> new ResourceNotFoundException("Resource not found for id: " + id));;
         System.out.println("product: " + product.getName());
         //ProductDto dto = new ProductDto(product.getId(), product.getName(), product.getDescription(), product.getPrice(), product.getImgUrl());
         ProductDto dto = new ProductDto(product);
@@ -52,14 +57,15 @@ public class ProductService {
 
     @Transactional
     public ProductDto update(Long id, ProductDto dto) {
-
-        Product entity = productRepository.findById(id).get();
-
-        copyDtoToEntity(dto, entity);
-
-        entity = productRepository.save(entity);
-        System.out.println("entity updated: " + entity.getName());
-        return new ProductDto(entity);
+        try {
+            Product entity = productRepository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = productRepository.save(entity);
+            System.out.println("entity updated: " + entity.getName());
+            return new ProductDto(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Resource not found for id: " + id);
+        }
     }
 
     private void copyDtoToEntity(ProductDto dto, Product entity) {
@@ -69,9 +75,16 @@ public class ProductService {
         entity.setImgUrl(dto.getImgUrl());
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not found for id: " + id);
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("referential integrity failure");
+        }
     }
 
 }
